Add files via upload
Write and run a program that calculates and displays the amount of money available in a bank 
account that initially has $1,000 deposited in it and that earns 8 percent interest a year. Your 
program should display the amount available at the end of each year for a period of ten years. Use 
the relationship that the money available at the end of each year equals the amount of money in 
the account at the start of the year plus .08 times the amount available at the start of the year. 
Your program should complete and produce a display as follows with values beneath the titles.
Year          Amount(beginning Yr)                 Amount(end Yr)